requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.7
Django==4.2.4
envparse==0.2.0
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
